requirements.txt
ipykernel
ipywidgets
python-dotenv
groq
langchain_groq
jupyter
streamlit
fastapi